Address comments and update readme
